seperate uniforms from shader file add uniformdata class
